feat: use location.message if available as the file name
Having "Primary Location" display the same value for every row (such as `/package.json` or `/pom.xml`) is not very user friendly.

SARIF allows a "message" to be provided for a location and that can have
more useful information. See:
https://docs.oasis-open.org/sarif/sarif/v2.1.0/os/sarif-v2.1.0-os.html#_Toc34317675

The result is that instead of showing many rows with "Primary Location"
of "/pom.xml", now each row has a different value such as "Package:
org.postgresql:postgresql@42.3.6" and "package: org.yaml:snakeyaml@1.30"

Signed-off-by: Craig Andrews <candrews@integralblue.com>